Attempts to determine display issue
Added text to dalek model to display tag at least.
Added script to display console but I don't have great faith the console is providing useful information.
Need to see XHR to know models have loaded.
The biggest file is the audio file and it is downloading AND playing. Not sure what is happening with models especially the Dalek.