Dalek removed edgesplit mod
The only objects not to appear in the previous commit were those with the edgesplit modifier applied. This is inconsistant but is the only clear difference in the various meshes.
NOTE: When exporting 'Apply Modifiers' was checked. So... not sure what's happening unless it is the splitting of the mesh that has caused the issue. Let's see what happens.