Dalek Materials removed
GLTF does not support node based texture correctly from Blender. They render in desktop browser but not on Safari iOS. Materials don't render so the model is not rendered. Sucks. This upload I've removed all but the materials on the dome lights to see what happens. If they are not displayed which is what I suspect I will remove the materials and upload the model 'bare'.