HTML CSS updates
Repsonsiveness has been added to both Dalek and main Github page. Added a standin character to Dalek file. Each represents the correct height for 3 main characters: Ian Chesterton on the left, Susan and The Doctor in the center. Barbara is the same height as The Doctor. This gives a sense of the room scale.