Skaro Issues
Having trouble getting the scene to display on an iPhone. Changes have removed all geometry so not sure what is going on. Removed Standin figures to see if that was the problem but no. Still nothing displayed. Removed dalek model and updated js to reflect lack of models except room. Commented out loaded images.