Pot/fill light update
Removed Ambient light as it was not affecting the set just the primative.
Removed all but the corner pots and brightened them up. There looks to be a limit on the number of lights allowed in the scene (?). There are now just four.
Counting the ambient light, 5 lights have been removed.