Refresh: Grid mkup fixed and reorg or files
Cleaned up grid markup on main page. Added a new AR hit test file to replace the old one which crashed a browser (?!) and a scene file for testing with HMDs. This scene Aframe_Level is just barebones and does not yet have the controllers enabled. It does have a nav-mesh though.