Collection done & minor clean up
Finally detected when the collpos animation was completed and hid the coin. There are issues with getting it to move to the correct location but it appears I missed the closing bracket on the camRig entity. That seem to be the problem but there are still some weird things going on. The camera and rig get out of position sync which I don't understand. Anyway, this works well enough to move on.

Did a little clean up to comment out console.logs I don't need to fire off right now. Will put updates into older test files.