grabbingtest moved
To match the working version of grab in VR_basic_teleport_collect I've moved the grabbingtest to the objects that are collectable. Point of fact, I've moved it to the mixin so we'll see if that works.

This may fix the listeners used for grab-start and hover-start. I've also placed some code to disable the raycaster on the cursor but have commented it out for now. Will use it if necessary. It appears the working test file mentioned above only had the hidden set to false and it stopped it from working.