Game component complete
Added 'game play' in to the file so when player collects all the coins, the fuel gem appears and when that is collected, the submarine is available to click to end game.

Next steps: UI, sound and loading screens. Will also check to see how the end game camera (cineCam) works with HMD and this may need further exploring.