Added Submarine Animation - propellor
Updated mesh to include 'steps' and round corners in the navmesh.
Submarine colour updated and propellor animation added. This has been implemented in code and index file.
Tried implementing all of aframe extras but it had an error so just added the loaders components which have the animation mixer.
new version is 1.3.5