Small adjustments to 3 projects
Basic_ui: the ambient light does not affect the town gltf. Other lights do with the exception of hemisphere. The ambient light does affect the primitives in the scene,
Animated_gltf: Added an ambient light to scene and it does affect the character gltf meshes in the scene. I did this to make sure the error I'm observing in basic_ui is only (currently) affecting the town model.
Submariner: just added positional awareness for the audio on the sub.